Revert "fixed: remove unused detect"
This reverts commit ff6e7969e8a15a02ba2656ac50595f04435b7b61.

+++ install.sh
@@ -120,7 +120,6 @@ echo ""
   # Installing zsh plugins
   echo ""
   echo "Installing zsh plugins ......"
-  echo ""
   rm -rf $HOME/.oh-my-zsh/custom/plugins/zsh-autosuggestions
   git clone https://github.com/zsh-users/zsh-autosuggestions.git $HOME/.oh-my-zsh/custom/plugins/zsh-autosuggestions
   echo ""
@@ -149,7 +148,9 @@ echo ""
   # Copy .zshrc
   echo ""
   echo "Copying .zshrc into home directory......"
-  rm $HOME/.zshrc
+  if [ -f "${$HOME}/.zshrc" ]; then
+    rm $HOME/.zshrc
+  fi
   cp .zshrc $HOME/
   echo "🏆 Copy .zshrc to home directory successfully!"
   echo ""
@@ -158,7 +159,9 @@ echo ""
   # Copy .tmux.conf.local
   echo ""
   echo "Copying .tmux.conf.local into home directory ......"
-  rm $HOME/.tmux.conf.local
+  if [ -f "${$HOME}/.tmux.conf.local" ]; then
+    rm $HOME/.tmux.conf.local
+  fi
   cp .tmux.conf.local $HOME/
   echo "🏆 Copy .tmux.conf.local to home directory successfully!"
   echo ""
@@ -184,7 +187,6 @@ echo ""
   if [ ! -d "${$HOME}/.config/nvim" ]; then
     mkdir $HOME/.config/nvim
   fi
-  rm $HOME/.config/nvim/init.vim
   cp .config/nvim/init.vim $HOME/.config/nvim/init.vim
   echo "🏆 Copy init.vim to ~/.config/nvim/init.vim successfully!"
   echo ""
